refactor(timeline): extract scroll-tracking effects into named hooks
Both TimelineOverview and TimelineToolbar buried a rAF-throttled
scroll/resize-tracking useEffect in their component body, making it
hard to tell what the component does without reading the whole
effect. Pull each into its own hook named for what it returns:
useTimelineViewportSize (the overview's viewport window) and
useScrollEdgeFade (the day-row's overflow fade) - so the component
body reads as a short list of "what this does" instead of inline
implementation.
